Move Core\Config to Config\Config
