+++ app/Config/Languages.php
@@ -7,7 +7,7 @@
  * @version 3.0
  */
 
-use Core\Config;
+use Config\Config;
 
 
 Config::set('languages', array(
