+++ app/Config/Modules.php
@@ -7,7 +7,7 @@
  * @version 3.0
  */
 
-use Core\Config;
+use Config\Config;
 
 
 Config::set('modules', array(
