+++ app/Modules/Users/Config.php
@@ -7,7 +7,7 @@
  * @version 3.0
  */
 
-use Core\Config;
+use Config\Config;
 
 /**
  * Configuration constants and options.
