+++ system/Cli/ClearSessionsCommand.php
@@ -7,7 +7,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-use Core\Config;
+use Config\Config;
 
 class ClearSessionsCommand extends Command
 {
